Refactor tests: move integration test to separate file, remove unused tests
